PlatformCameraComponent, CameraBounds
~ Finished implementing layer's

+++ Source/Grief/CameraBoundingBox.h
@@ -38,11 +38,11 @@ public:
 	bool IsBoundActive(EDirection Direction);
 
 	UFUNCTION(BlueprintPure)
-	FORCEINLINE uint8 GetLayer() const { return Layer; }
+	FORCEINLINE int32 GetLayer() const { return Layer; }
 	
 private:
 	UPROPERTY(EditInstanceOnly, BlueprintReadOnly, Category = "Options", meta = (AllowPrivateAccess = "true"))
-	uint8 Layer = 0;
+	int32 Layer = 0;
 	
 	UPROPERTY(EditInstanceOnly, BlueprintReadOnly, Category = "Options|Bounds", meta = (AllowPrivateAccess = "true"))
 	bool Up = true;
